actualizacion de clase Program.cs

+++ ServicioSOA/ServicioSOA/Modelo/Producto.cs
@@ -5,11 +5,11 @@
         //atributos
         public int id {get;set;}
         public string nombre {get;set;}
-        public decimal pvp {get;set;}
+        public double pvp {get;set;}
         public int cantidad {get;set;}
 
         //constructor
-        public Producto(int id, string nombre, decimal pvp, int cantidad)
+        public Producto(int id, string nombre, double pvp, int cantidad)
         {
             this.id = id;
             this.nombre = nombre;
